+++ amd/src/player/sproutvideo.js
@@ -389,6 +389,7 @@ class SproutVideo {
      */
     setCaption() {
         // Not supported
+        return false;
     }
 
     /**